requirements.txt
numpy
Django==5.0.2
python-dotenv==1.0.1
Requests==2.31.0
schedule==1.2.1
django-allauth==0.61.1
fake_useragent
django-cors-headers
pillow
# psycopg2
psycopg2-binary
beautifulsoup4==4.12.3
huggingface_hub
datasets
GitPython
sib_api_v3_sdk